% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_fb_id_names.R
\name{fill_fb_id_names}
\alias{fill_fb_id_names}
\title{Fill Names of IDs}
\usage{
fill_fb_id_names(fb_df, version, token)
}
\arguments{
\item{fb_df}{Facebook dataframe, returned by \code{query_fb_marketing_api()}}

\item{version}{Facebook Marketing API version; for example, "v14.0"}

\item{token}{Facebook Marketing API token}
}
\value{
Dataframe with Facebook results, where additional variables are created that indicate the name of IDs
}
\description{
The Facebook data contains variables with IDs, where it can be useful to have
the name of the associated ID. This function creates variables that contain the
the name. In cases where there are multiple IDs, the names are separated by a
semicolon. For example, if there is an \code{interest} variable with the
observation: \code{"6003349442621,6003020834693"}, an \code{interest_name} variable
with "Entertainment; Music" would be created.
}
\examples{
\dontrun{
fb_df <- query_fb_marketing_api(location_unit_type = "coordinates",
                                lat_lon = c(38.90, -77.01),
                                radius = 10,
                                radius_unit = "kilometer",
                                interest = list(c(6002839660079,
                                                  6002884511422),
                                                6002839660079,
                                                6002884511422),
                                behavior = list(c(6002714895372,
                                                  6002714898572),
                                                6002714895372,
                                                6002714898572),
                                version = VERSION, 
                                creation_act = CREATION_ACT, 
                                token = TOKEN)

fb_df <- fill_fb_id_names(fb_df,
                         version = VERSION,
                         token = TOKEN)
}
}
