% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_fb_parameters}
\alias{get_fb_parameters}
\title{Get Facebook Parameter IDs}
\usage{
get_fb_parameters(type, version, token)
}
\arguments{
\item{type}{Type of data; either "behaviors", "demographics", "interests", or "locales"}

\item{version}{Facebook Marketing API version; for example, "v14.0"}

\item{token}{Facebook Marketing API token}
}
\value{
Dataframe with parameter IDs and descriptions.
}
\description{
This function returns dataframes of parameters for behaviors, demographics,
and interests. The dataframes contain ids that can be used in the
query_fb_marketing_api function.
}
