% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fb_marketing_api.R
\name{map_param_vec}
\alias{map_param_vec}
\title{Map Parameters over Vector
Instruct \code{query_fb_marketing_api()} to make separate queries for each item in a vector.}
\usage{
map_param_vec(...)
}
\arguments{
\item{...}{Vector}
}
\description{
Map Parameters over Vector
Instruct \code{query_fb_marketing_api()} to make separate queries for each item in a vector.
}
\examples{
\dontrun{
# Make 2 queries:
# 1. Number of male MAU/DAU
# 2. Number of female MAU/DAU
query_fb_marketing_api(
  location_unit_type = "countries",
  location_keys      = "US",
  gender             = map_param(1:2)
  version            = VERSION, 
  creation_act       = CREATION_ACT, 
  token              = TOKEN)
}
}
