% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_fb_id_names.R
\name{merge_fb_with_param_df}
\alias{merge_fb_with_param_df}
\title{Merge Facebook Data with Parameter Name}
\usage{
merge_fb_with_param_df(fb_df, param_df, var)
}
\arguments{
\item{fb_df}{Facebook dataframe, returned by \code{query_fb_marketing_api()}}

\item{param_df}{Parameter dataframe, returned by \code{get_fb_parameter_ids()}}

\item{var}{Parameter variable name (in \code{fb_df})}
}
\value{
Facebook dataframe, with name of variable
}
\description{
The Facebook data will have an ID for a number of variables, where it is also
useful to have the name associated with the ID. Using a dataframe of Facebook
results and a parameter dataframe (that contains the names and ids), add
the names to the Facebook data. Deals with cases where there are multiple
ids used.
}
