% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fb_parameter_ids.R
\name{get_fb_parameter_ids}
\alias{get_fb_parameter_ids}
\title{Get Facebook Parameter IDs}
\usage{
get_fb_parameter_ids(
  type,
  version,
  token,
  q = NULL,
  country_code = NULL,
  region_id = NULL,
  key = NULL
)
}
\arguments{
\item{type}{Type of data; either "behaviors", "demographics", "interests", or "locales"}

\item{version}{Facebook Marketing API version; for example, "v14.0"}

\item{token}{Facebook Marketing API token}
}
\value{
Dataframe with parameter IDs and descriptions.
}
\description{
This function returns dataframes of Facebook parameters and their associated
IDs for different categories of information. Categories include behaviors,
demographics, interests, locales, job titles, education major, and location
(e.g., country, city, zip code, etc). The returned dataframe contains ids
that can be used in the query_fb_marketing_api function.
}
