% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_param.R
\name{map_param}
\alias{map_param}
\title{Map Parameters
Converts a vector or list into a \code{map_param} object. When a \code{map_param} object is entered as a parameter in \code{query_fb_marketing_api()}, \code{query_fb_marketing_api()} makes a separate query for each item within \code{map_param}.}
\usage{
map_param(...)
}
\arguments{
\item{...}{Vector or list}
}
\value{
Object of class \code{map_param}; a list structured that contains different parameters for the \code{query_fb_marketing_api()} function.
}
\description{
Map Parameters
Converts a vector or list into a \code{map_param} object. When a \code{map_param} object is entered as a parameter in \code{query_fb_marketing_api()}, \code{query_fb_marketing_api()} makes a separate query for each item within \code{map_param}.
}
\examples{
\dontrun{
# Make 3 queries:
# 1. Number of males and females MAU/DAU
# 2. Number of male MAU/DAU
# 3. Number of female MAU/DAU
query_fb_marketing_api(
  location_unit_type = "countries",
  location_keys      = "US",
  gender             = map_param(c(1,2), 1, 2),
  version            = VERSION, 
  creation_act       = CREATION_ACT, 
  token              = TOKEN)
}
}
