% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{query_fb_marketing_api}
\alias{query_fb_marketing_api}
\title{Query Facebook Marketing API}
\usage{
query_fb_marketing_api(
  location_type,
  lat_lon = NULL,
  radius = NULL,
  radius_unit = NULL,
  country_iso2 = NULL,
  education_statuses = NULL,
  user_os = NULL,
  wireless_carrier = NULL,
  behavior = NULL,
  interest = NULL,
  gender = c(1, 2),
  age_min = 18,
  age_max = 65,
  sleep_time = 1,
  show_result = T,
  version,
  creation_act,
  token,
  add_query = F,
  add_query_hide_credentials = T
)
}
\arguments{
\item{location_type:}{Either "coordinates" (for buffer around single point) or "country"}
}
\description{
Query Facebook Marketing API
}
